snapshot: clonetree: handle PCI bridges
Linux does not flatten the PCI tree in sysfs.
This is a fine and valid choice, but which is easy to overlook
when exploring simple PCI device trees, such as ones found on laptops,
but becomes evident when exploring more complex trees, like the ones
we can see on servers, even low end.

This means, in case of PCI bridges, we will have
entries inside the directory which represent a bus which
will be both devices and root of subtrees.

In other words, these entries will contain BOTH subdirectories
to represent the devices attached to that bridge, and the files
which describe the bridge device per se.

For ghw, the net effect is we cannot just scan the entries
in `/sys/bus/pci/devices` and be done with it; we need to detect
the entries which represent PCI briges, and scan those recursively.
Otherwise, we miss to gather the data pertaining to these devices.

So we need a smarter function that does this traversal, moving away
from the nice static list of paths we previously had.

Signed-off-by: Francesco Romani <fromani@redhat.com>